test(integration): remove suites duplicating existing coverage
The vsr flavor made several suites triple-cover the same command
surface: sdk/mcp_parity.rs mirrored the mcp::* suite, which is not
feature-gated and already runs against server-ng; sdk/messages.rs
repeated a system_scenario subset; create_message_payload was a
clone of message_headers_scenario differing only in constants.

Fold the solo-replica checkpoint recovery test into
cluster/metadata_checkpoint_restart.rs next to the other checkpoint
boundary cases. Drop the permanently ignored sdk_e2e test in
server-ng: a default-features SDK cannot speak to the ng listener,
and the covered flow is exercised by
cluster::multi_shard_partition_convergence under vsr. The bind
smoke test stays as the only default-lane process-level server-ng
test.

With the e2e gone server-ng no longer needs the iggy SDK dev-dep,
so its vsr feature forwards only iggy_common/vsr; every consumer
lane (cli, integration, bdd) carries its own iggy/vsr forward.
