+++ core/integration/tests/sdk/mod.rs
@@ -21,10 +21,6 @@ mod consumer_group_membership;
 mod hello_world;
 #[cfg(feature = "vsr")]
 mod http_refresh;
-#[cfg(feature = "vsr")]
-mod mcp_parity;
-#[cfg(feature = "vsr")]
-mod messages;
 mod producer;
 #[cfg(feature = "vsr")]
 mod protocol_version;
