+++ core/integration/tests/server/mod.rs
@@ -48,10 +48,6 @@ mod cluster_metadata_vsr;
 // across a replica restart.
 #[cfg(feature = "vsr")]
 mod cluster_view_durability_vsr;
-// A metadata checkpoint must drain the WAL and recover from the snapshot fold plus
-// the WAL suffix across a restart.
-#[cfg(feature = "vsr")]
-mod metadata_checkpoint_recovery_vsr;
 // 80-case race matrix with hardcoded HTTP variants (test_matrix bypasses
 // the harness transport filter).
 mod concurrent_addition;
